wirego_remote(python): create initial pyproject with defined dependencies
