+++ .github/workflows/python-ubuntu.yml
@@ -1,0 +1,47 @@
+name: Python Wirego Remote build on Linux
+
+on:
+  push:
+    branches: ["main"]
+    paths: [".github/workflows/python-ubuntu.yml", "wirego_remote/python/**"]
+  pull_request:
+    branches: ["main"]
+    paths: [".github/workflows/python-ubuntu.yml", "wirego_remote/python/**"]
+
+jobs:
+  linux:
+    name: Build
+    runs-on: ubuntu-latest
+    env:
+      working-directory: ${{ github.workspace }}/wirego_remote/python/wirego
+    strategy:
+      matrix:
+        python_version: ["3.9", "3.10", "3.11", "3.12", "3.13", "3.14"]
+        poetry_version: ["2.1.3"]
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python_version }}
+
+      - name: Install Poetry
+        uses: abatilo/actions-poetry@v4
+        with:
+          poetry-version: ${{ matrix.poetry_version }}
+
+      - name: Install poetry dependencies
+        run: poetry install
+        working-directory: ${{ env.working-directory }}
+
+      - name: Run formatter check
+        run: poetry run ruff format --check
+        working-directory: ${{ env.working-directory }}
+
+      - name: Run linter check
+        run: poetry run ruff check
+        working-directory: ${{ env.working-directory }}
+
+      - name: Run tests
+        run: echo "Running tests is not implemented yet"
+        working-directory: ${{ env.working-directory }}
