+++ wirego_remote/python/wirego/examples/minimal/screenshot.png
