+++ wirego_remote/python/wirego/examples/minimal/wirego_minimal.pcap
