+++ wirego_remote/python/wirego/src/wirego/__init__.py
@@ -1,0 +1,13 @@
+from .wirego import (
+    ValueType,
+    DisplayMode,
+    DetectionFilterType,
+    WiregoField,
+    DetectionFilter,
+    DissectField,
+    DissectResult,
+    DissectResultFieldFlatten,
+    DissectResultFlattenEntry,
+    WiregoListener,
+    Wirego,
+)
